Add files via upload
Tariffs are Taxes, Plain and Simple. This handy calculator lets Americans figure out how much the largest tax increase in history will hit them where it hurts. In the balls. And the spatulas. And all the other cheap crap we buy from China and think we are going to start manufacturing here in America.